rename some of the files name to make more sense
