+++ src/lib/app-watcher.js
